v102/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-tables2==2.7.5
et_xmlfile==2.0.0
Faker==24.2.0
gunicorn==23.0.0
idna==3.10
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==8.2
reportlab==4.4.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
xlsxwriter==3.2.5
